requirements.txt
# Sleep Video Generator Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
openai>=1.0.0
moviepy>=1.0.3
pillow>=9.0.0
python-dotenv>=1.0.0
requests>=2.28.0

# Optional: For better video processing
imageio>=2.22.0
imageio-ffmpeg>=0.4.7

# Optional: For better font handling (system dependent)
# On macOS/Linux: Usually included with system
# On Windows: May need additional font packages